LibWeb: Share the retained fly-string types across generated headers
RetainedUtf16FlyString and its list type lived in style_value.rs, whose
cbindgen invocation is the only one that parses them, so a style group
payload in the computed-values header could not embed the list without
duplicating its whole unsafe retain/release lifecycle. Move the two
types into a types-only module parsed by both the style-value and
computed-values header invocations — the arrangement display.rs already
uses for FfiDisplay — and give the list a Clone impl so payload structs
can hold it. The struct literals that built fly strings from leaked raws
inside style_value.rs now spell out from_leaked_raw, since the raw field
is no longer visible there.
